+++ R/market_model.R
@@ -267,10 +267,10 @@ setMethod("show", signature(object = "market_model"), function(object) {
 #' Summarizes the  model.
 #'
 #' Prints basic information about the passed model object. In addition to the output of
-#' the \code\link{{show}} method, \code{summary} prints
+#' the \code{\link{show}} method, \code{summary} prints
 #' - the number of observations,
 #' - the number of observations in each equation for models with sample separation, and
-#' - the number of (various categories of) variables.
+#' - various categories of variables.
 #' @param object A model object.
 #' @examples
 #' \donttest{