+++ R/system_stochastic_adjustment.R
@@ -117,10 +117,13 @@ setMethod(
   }
 )
 
-setMethod("show", signature(object = "system_stochastic_adjustment"), function(object) {
-    callNextMethod(object)
-    show(object@price_equation)
-})
+setMethod(
+    "show_implementation", signature(object = "system_stochastic_adjustment"),
+    function(object) {
+        callNextMethod(object)
+        show_implementation(object@price_equation)
+    }
+)
 
 setMethod(
   "get_likelihood_variables", signature(object = "system_stochastic_adjustment"),